+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/commands.py
@@ -69,6 +69,8 @@ def load_command_table(self, _):
                          confirmation='Kubernetes may be unavailable during cluster upgrades.\n' +
                          'Are you sure you want to perform this operation?')
         g.custom_command('upgrade-connector', 'k8s_upgrade_connector')
+        g.custom_command('use-dev-spaces', 'aks_use_dev_spaces')
+        g.custom_command('remove-dev-spaces', 'aks_remove_dev_spaces')
         g.generic_wait_command('wait')
 
     with self.command_group('aks', container_services_sdk, client_factory=cf_container_services) as g: