+++ src/azure-cli-testsdk/HISTORY.rst
@@ -3,6 +3,14 @@
 Release History
 ===============
 
+0.1.3
++++++
+* Update azure-devtools to 1.0.0
+
+0.1.2
++++++
+* Update the setup.py to limit the dependency on azure-devtools in 0.5.* range.
+
 0.1.1
 +++++
 * Add additional tags to the resource group created by automation