+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -2,6 +2,11 @@
 
 Release History
 ===============
+0.1.33
+++++++
+* webapp/functionapp: improve generic update commands
+* webapp/functionapp: webapp deployment source config-zip supports async operation with status updates for long running operation 
+
 0.1.32
 ++++++
 * webapp: fix a bug in `az webapp delete` when `--slot` is provided