+++ src/command_modules/azure-cli-extension/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+0.0.14
+++++++
+* Be more resilient to system error when removing an extension.
+
 0.0.13
 ++++++
 * Pin version of `wheel` so extensions can get metadata shown again.