+++ src/command_modules/azure-cli-keyvault/HISTORY.rst
@@ -2,10 +2,14 @@
 
 Release History
 ===============
-2.0.22
-++++++
+Current
++++++++
 * adding commands for managing storage accounts and sas-definitions
 * adding commands for network-rules
+
+2.0.22
+++++++
+* make keyvault commands work in cloud shell or vms with identity
 
 2.0.21
 ++++++